Moved scripts, cleanup
