+++ Assets/_Functionalities/WFS/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e24cc56eac6e54b9991ee617fb533653
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
