+++ Assets/_Functionalities/WFS/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: dcfcafe22934e4fae9399fe31a99479d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
