+++ Assets/_Functionalities/WFS/Scripts/WFSGeoJSONLayerGameObject.cs
@@ -2,8 +2,10 @@ using UnityEngine;
 using Netherlands3D.Twin.Layers.Properties;
 using System.Collections.Generic;
 using System.Linq;
+using Netherlands3D.Twin.Layers.LayerTypes.GeoJsonLayers;
+using Netherlands3D.Twin.Utility;
 
-namespace Netherlands3D.Twin.Layers.LayerTypes.GeoJsonLayers
+namespace Netherlands3D.Functionalities.Wfs
 {
     /// <summary>
     /// Extention of GeoJSONLayerGameObject that injects a 'streaming' dataprovider WFSGeoJSONTileDataLayer
