+++ Assets/_Functionalities/Wms/Scripts/WMSImportAdapter.cs
@@ -1,11 +1,9 @@
 using System;
-using System.Collections;
 using System.IO;
 using Netherlands3D.DataTypeAdapters;
 using Netherlands3D.Twin.Layers.LayerTypes;
 using Netherlands3D.Twin.Layers.Properties;
 using UnityEngine;
-using UnityEngine.Networking;
 
 namespace Netherlands3D.Functionalities.Wms
 {
